Move to core-api UPayload data model
Pulling from snapshot till the other project is merged

#40
