+++ src/main/java/org/eclipse/uprotocol/rpc/RpcClient.java
@@ -26,7 +26,7 @@ package org.eclipse.uprotocol.rpc;
 
 import java.util.concurrent.CompletableFuture;
 
-import org.eclipse.uprotocol.transport.datamodel.UPayload;
+import org.eclipse.uprotocol.v1.UPayload;
 import org.eclipse.uprotocol.v1.UAttributes;
 import org.eclipse.uprotocol.v1.UUri;
