+++ src/main/java/org/eclipse/uprotocol/rpc/RpcMapper.java
@@ -33,7 +33,7 @@ import com.google.rpc.Status;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 
-import org.eclipse.uprotocol.transport.datamodel.UPayload;
+import org.eclipse.uprotocol.v1.UPayload;
 
 /**
  * RPC Wrapper is an interface that provides static methods to be able to wrap an RPC request with 
@@ -60,7 +60,7 @@ public interface RpcMapper {
             }
             Any any;
             try {
-                any = Any.parseFrom(payload.data());
+                any = Any.parseFrom(payload.getValue());
             
                 // Expected type
                 if (any.is(expectedClazz)) {
@@ -93,7 +93,7 @@ public interface RpcMapper {
             }
             Any any;
             try {
-                any = Any.parseFrom(payload.data());
+                any = Any.parseFrom(payload.getValue());
                 
                 // Expected type
                 if (any.is(expectedClazz)) {
