+++ src/main/java/org/eclipse/uprotocol/transport/datamodel/UListener.java
@@ -2,6 +2,7 @@ package org.eclipse.uprotocol.transport.datamodel;
 
 import org.eclipse.uprotocol.v1.UAttributes;
 import org.eclipse.uprotocol.v1.UUri;
+import org.eclipse.uprotocol.v1.UPayload;
 
 /**
  * For any implementation that defines some kind of callback or function that will be called to handle incoming messages.
